melanoma-backend-main/melanoma-backend-main/requirements.txt
anyio==4.11.0
asgiref==3.10.0
asttokens==2.4.1
bleach==6.3.0
blinker==1.8.2
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.1.7
colorama==0.4.6
comm==0.2.2
debugpy==1.8.2
decorator==5.1.1
Django==5.2.8
django-cors-headers==4.9.0
django-rest-framework==0.1.0
djangorestframework==3.16.1
executing==2.0.1
filelock==3.20.0
Flask==3.0.3
fsspec==2025.10.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.1.4
idna==3.11
ipykernel==6.29.5
ipython==8.26.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
kaggle==1.7.4.5
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.5
nltk==3.8.1
numpy==1.26.4
packaging==24.1
pandas==2.3.3
parso==0.8.4
pillow==12.0.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
protobuf==6.33.0
psutil==6.0.0
pure_eval==0.2.3
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2025.2
pywin32==306
PyYAML==6.0.3
pyzmq==26.0.3
regex==2024.7.24
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.5.1
scipy==1.14.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.3
stack-data==0.6.3
sympy==1.14.0
text-unidecode==1.3
threadpoolctl==3.5.0
timm==1.0.22
torch==2.9.1
torchvision==0.24.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typer-slim==0.20.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.3
whitenoise==6.11.0
